test: split error test cases into serialization/business logic
